requirements.txt
# Python依赖管理
# 基础框架
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-multipart==0.0.6

# 数据库
sqlalchemy==2.0.23
aiosqlite==0.19.0
alembic==1.12.1

# YAML处理
pyyaml==6.0.1
ruamel.yaml==0.18.5

# 工作流引擎
asyncio==3.4.3
aiohttp==3.9.1

# 云平台
kubernetes==28.1.0

# 工具
python-jose==3.3.0
passlib==1.7.4
python-dateutil==2.8.2
aiofiles==23.2.1

# 测试工具
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# 类型提示
typing-extensions==4.8.0 